pyproject.toml
[tool.poetry]
name = "gap-mm"
version = "0.1.0"
description = "Gap-probability market maker for Bybit (Rust + Python + Numba)"
authors = ["ali askar <26202651+alihaskar@users.noreply.github.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/alihaskar/gap-mm"
repository = "https://github.com/alihaskar/gap-mm"
packages = [{ include = "gap_mm", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
numba = "^0.63.1"
python-dotenv = "^1.0.0"
requests = "^2.32.5"

[tool.poetry.group.dev.dependencies]
maturin = "^1.11.5"
pytest = "^8.3"
pytest-mock = "^3.14"
ruff = "^0.11"
pytest-httpserver = "^1.1.5"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["gap_mm"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
